+++ package.json
@@ -15,6 +15,7 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.469.0",
+    "material-symbols": "^0.27.2",
     "next": "15.1.2",
     "next-themes": "^0.4.4",
     "react": "^19.0.0",
