+++ src/app/layout.js
@@ -1,21 +1,10 @@
-import { ThemeProvider } from "../components/ui/themeProvider"
-// import { NavigationBar } from "@/components/navigation-bar"
-import Navbar from "../components/ui/navbar"
+import RootLayout from "./rootLayout"
+import TypographyH2 from "@/components/ui/typography"
 
-export default function RootLayout({ children }) {
+export default function Layout({ children }) {
     return (
-        <html lang="en" suppressHydrationWarning>
-            <body>
-                <ThemeProvider
-                    attribute="class"
-                    defaultTheme="system"
-                    enableSystem
-                    disableTransitionOnChange
-                >
-                    <Navbar />
-                    {children}
-                </ThemeProvider>
-            </body>
-        </html>
+        <RootLayout>
+            <TypographyH2 />
+        </RootLayout>
     )
 }
