+++ src/components/ui/typography.jsx
@@ -1,3 +1,4 @@
+"use client"
 import React from "react"
 
 const TypographyH2 = () => {
